Fix bug #2
Reason :
The bug was getting caused by the fact that cursor is made using the
left-border property of a word, which makes words a little bit thicker
when the cursor is applied, but when the cursor goes to the next word
, the previous word's thickness lessens by the cursor's width.

Due to this behaviour, when the cursor goes to the next line's second
word, sometimes the first word's width becomes as less as was needed
to get fit in the first line.

Fix :
Including a padding equal to the cursor's width in the TypingArea fixes
this issue, as now the cursor's width will no longer be of a concern.