Feature : one word gets added after compeletion of typing one word.
As soon as user types the whole word, a word is added to the end
of the typingContent.

This feature enables us to not load and store a huge amount of
typingContent in any given time.

We can now add words on the fly(after compeletion of each word),
which enables us to have only a small amount of typingContent in memory.